Added course module with components func
