+++ core/data/calendar.py
@@ -24,6 +24,22 @@ class Calendar(Request):
          return sub(compile('<.*?>'), '', value)
 
 
+    def filter(self, value=None, data=None, *args, **kwargs):
+        filtering = 'Aula ao vivo - BigBlueButton' if value == 'bbb' else 'Tarefa para entregar via Moodle'
+
+        if value and data:
+            filtered = []
+
+            for row in data:
+                if row[3] == filtering:
+                    filtered.append(row)
+
+            return filtered
+
+        else:
+            raise ValueError('Value or data parameter not passed correctly.')
+
+
     def monthly(self, month=str(dt.today().month), year=str(dt.today().year), *args, **kwargs):
         month = Request.get(self, function='core_calendar_get_calendar_monthly_view', year=year, month=month)['weeks']
 
@@ -32,7 +48,8 @@ class Calendar(Request):
         for week in month:
             for day in week['days']:
                 for events in day['events']:
-                    if events['modulename'] == 'assign' or events['modulename'] == 'bigbluebuttonbn':
+                    if (events['modulename'] == 'assign' or events['modulename'] == 'bigbluebuttonbn')\
+                            and events['course']['fullname'] != 'ALGORITMOS E PROGRAMACAO I [turma 01G] - 2020/2':
                         data.append(
                             [
                                 events['course']['fullname'],
@@ -50,6 +67,8 @@ class Calendar(Request):
                                 + events['formattedtime'].split(':')[1][0:2],
 
                                 events['url'],
+
+
                             ]
                         )
