+++ csvfiles/events.csv
@@ -1,4 +1,3 @@
-ALGORITMOS E PROGRAMACAO I [turma 01G] - 2020/2,BBB - Variáveis e tipos de dados is scheduled for,Descrição não disponível,Aula ao vivo - BigBlueButton,"terça, 1 setembro",08:20,https://eadmoodle.mackenzie.br/mod/bigbluebuttonbn/view.php?id=885736
 INTRODUCAO A COSMOVISAO REFORMADA [turma 02D] - 2020/2,Catálogo Deism_Naturalis - Aula síncrona is scheduled for,Descrição não disponível,Aula ao vivo - BigBlueButton,"terça, 1 setembro",09:20,https://eadmoodle.mackenzie.br/mod/bigbluebuttonbn/view.php?id=888343
 ANALISE DE DADOS [turma 02D11] - 2020/2,Análise de Dados – 2D Lab 11 – Semana de 31/08 a 06/09 is scheduled for,Descrição não disponível,Aula ao vivo - BigBlueButton,"quarta, 2 setembro",09:20,https://eadmoodle.mackenzie.br/mod/bigbluebuttonbn/view.php?id=891120
 ANALISE DE DADOS [turma 02D] - 2020/2,Análise de Dados – 2D – 03/09 is scheduled for,Descrição não disponível,Aula ao vivo - BigBlueButton,"quinta, 3 setembro",09:20,https://eadmoodle.mackenzie.br/mod/bigbluebuttonbn/view.php?id=893750
